Change namespacing for volunteering history components
