+++ app/views/candidate_interface/application_form/_review.html.erb
@@ -55,7 +55,7 @@
 <%= render(CandidateInterface::WorkHistoryReviewComponent.new(application_form: application_form, editable: editable, heading_level: 3, show_incomplete: !application_form.work_history_completed && editable, missing_error: missing_error)) %>
 
 <h2 class="govuk-heading-m govuk-!-font-size-27"><%= t('page_titles.volunteering.short') %></h2>
-<%= render(VolunteeringReviewComponent.new(application_form: application_form, editable: editable, heading_level: 3, show_incomplete: true, missing_error: missing_error)) %>
+<%= render(CandidateInterface::VolunteeringReviewComponent.new(application_form: application_form, editable: editable, heading_level: 3, show_incomplete: true, missing_error: missing_error)) %>
 
 <h2 class="govuk-heading-m govuk-!-font-size-27">Qualifications</h2>
