+++ app/views/candidate_interface/volunteering/review/show.html.erb
@@ -7,7 +7,7 @@
       <%= t('page_titles.volunteering.short') %>
     </h1>
 
-    <%= render(VolunteeringReviewComponent.new(application_form: @application_form)) %>
+    <%= render(CandidateInterface::VolunteeringReviewComponent.new(application_form: @application_form)) %>
 
     <%= govuk_button_link_to t('application_form.volunteering.another.button'), candidate_interface_new_volunteering_role_path, class: 'govuk-button--secondary' %>
