+++ spec/components/previews/provider_interface/volunteering_history_component_preview.rb
@@ -14,7 +14,7 @@ module ProviderInterface
 
     def render_component_for(application_form:)
       if !application_form.application_work_experiences.empty?
-        render ProviderInterface::VolunteeringHistoryComponent.new(application_form: application_form)
+        render VolunteeringHistoryComponent.new(application_form: application_form)
       else
         render template: 'support_interface/docs/missing_test_data'
       end
